+++ src/backend/serial/mod.rs
@@ -21,7 +21,7 @@
 use cfg_if::cfg_if;
 
 cfg_if! {
-    if #[cfg(curve25519_dalek_backend = "fiat")] {
+    if #[cfg(curve25519_dalek_serial = "fiat")] {
 
         #[cfg(curve25519_dalek_bits = "32")]
         pub mod fiat_u32;