+++ src/backend/vector/avx2/field.rs
@@ -43,7 +43,11 @@ const D_LANES64: u8 = 0b11_00_00_00;
 use core::ops::{Add, Mul, Neg};
 use packed_simd::{i32x8, u32x8, u64x4, IntoBits};
 
+#[cfg(curve25519_dalek_serial = "fiat")]
+use crate::backend::serial::fiat_u64::field::FieldElement51;
+#[cfg(not(curve25519_dalek_serial = "fiat"))]
 use crate::backend::serial::u64::field::FieldElement51;
+
 use crate::backend::vector::avx2::constants::{
     P_TIMES_16_HI, P_TIMES_16_LO, P_TIMES_2_HI, P_TIMES_2_LO,
 };